Drop postgis_config.h include from lwgeom_pg.h header.
Generally, package config files should _not_ be included
by package headers. In this specific case the aim is overriding
POSTGIS_DEBUG_LEVEL in implementation files for the sake of
enabling debugging of a single file.

Also part of this commit is addition of copyright header in
two files which were missing it (used git history to figure
assignments)

git-svn-id: http://svn.osgeo.org/postgis/trunk@9134 b70326c6-7e19-0410-871a-916f4a2858ee