Add an internal topology._st_mintolerance funtion
This function computes the minimum tolerance for a given geometry.
It works by computing the smallest floating point difference that
could be expressed on any vertex of the geometry.

Currently it's only used to workaround the ST_DWithin/ST_Within
inconsistency (#1625) but could probably be used everytime 0 is
passed as a tolerance, possibly fixing #1641 as well.

git-svn-id: http://svn.osgeo.org/postgis/trunk@9374 b70326c6-7e19-0410-871a-916f4a2858ee