Rework code from r9204 (bug #900: shp2pgsql: a switch to drop M from 4d imports) to fix bug #1710: shp2pgsql -t option is not working properly.
The original implementation set the has_z/has_m flags directly based upon the -t parameter; however this lost the information
as to how many dimensions were specified. Resolve this by specifying a new force_output configuration variable and setting
it to an enumerated constant, so that we can then implement a simple switch() to set both the output dimension and output
flags correctly.

git-svn-id: http://svn.osgeo.org/postgis/trunk@9548 b70326c6-7e19-0410-871a-916f4a2858ee