Fix access to random memory during ANALYZE
The bug was introduced with previous commit, see #2269

As part of this fix, the stawidth field of pg_statistics is
changed to represent the average _compressed_ size rather
than the average _uncompressed_ size of non-null geometry
values. This may have some consequence on the planner (not
known at time of commit).

git-svn-id: http://svn.osgeo.org/postgis/branches/2.0@11275 b70326c6-7e19-0410-871a-916f4a2858ee