Cleanup raster drop files
rtpostgis_drop.sql is used to drop obsoleted objects (after)
while rtpostgis_upgrade_cleanup.sql is used to remove things that
need be removed _before_ (incompatible CREATE OR REPLACE).

Ideally nothing would need to be removed _before_ if the
proc_upgrade script was made smarter to determine incompatible
signature changes (but this is for another time).

git-svn-id: http://svn.osgeo.org/postgis/branches/2.1@12594 b70326c6-7e19-0410-871a-916f4a2858ee