Add COSTs to single polygon accepting functions
Costs are based on a set of 5M polygons from a UK OSM extract, based
on comparing runtimes with those of bigint addition, subtraction,
modulo, division, squaring, and square root.

The tests were done with PostGIS 2.1.8, so 2.2.0 and later functions
are not yet costed.

Also adds comments where costs are guessed.

ref https://lists.osgeo.org/pipermail/postgis-devel/2016-May/025808.html