#3549, parallel query in PgSQL 9.6 This patch mostly just marks a large number of functions as "parallel safe" when compiled against 9.6. As noted in
http://blog.cleverelephant.ca/2016/03/parallel-postgis.html
http://blog.cleverelephant.ca/2016/03/parallel-postgis-joins.html

this alone will not necessarily be sufficient to cause
parallel processing to kick in. We will additionally need to 
bump up function costs across the board, and even then may
require some extra kicking of GUCs to force spatial query
plans into parallelism as good moments.

The _PARALLEL keyword is used in sql.in files to add
PARALLEL SAFE to function definitions if invoked in the 
presence of PgSQL 9.6. This works fine in new installs
but upgrades will need to be tested.


git-svn-id: http://svn.osgeo.org/postgis/trunk@14905 b70326c6-7e19-0410-871a-916f4a2858ee