#3418, KNN recheck in 9.5 fails with index returned tuples in wrong order
While we store float boxes, it is important to carry out *comparisons*
of those boxes in double space, so we can capture small differences. 
This works well because the boxes are "overdetermined", they are slightly
larger than they need to be. The bug in this case was caused by a case
where the distance calculation between two boxes was carried out in float
space and as a result the distance got slightly overdetermined, ending
up larger than the actual distance between the objects themselves.



git-svn-id: http://svn.osgeo.org/postgis/trunk@15283 b70326c6-7e19-0410-871a-916f4a2858ee