#3837, --enable-debug to set PARANOIA level
This required some re-working of the protobuf sections,
for reasons that aren't clear. However, hopefully that
re-working will also fix things for others. Basically,
the pkg-config portions seemed to only work by accident
for the folks who wrote them initially, so they are 
improved, and the use of --without is more strongly
enforced in the conditional blocks


git-svn-id: http://svn.osgeo.org/postgis/trunk@15662 b70326c6-7e19-0410-871a-916f4a2858ee