Proof of content for reading table extent from the top of the index. Can be called with _postgis_index_extent(tbl regclass, col text) and returns box2d.
Currently not added to the st_estimatedextent()
code, because it has issues with indexes that
have EMPTY added to them. This is a symptom 
of #2487. While it seems impossible to every
allow index-assisted searches for EMPTY 
geometry, it will be necessary to create
invalid keys for them (NaN, NaN, NaN, Nan)
and then handle those keys so they don't 
pollute the index with strange bounds.

References #2256


git-svn-id: http://svn.osgeo.org/postgis/trunk@16149 b70326c6-7e19-0410-871a-916f4a2858ee