MVT tests: Remove ST_Normalize calls and check the orientation
From the spec (https://git.io/vhrK4):
```
An exterior ring is DEFINED as a linear ring having a positive area
as calculated by applying the surveyor's formula to the vertices of
the polygon in tile coordinates. In the tile coordinate system
(with the Y axis positive down and X axis positive to the right)
this makes the exterior ring's winding order appear clockwise.

An interior ring is DEFINED as a linear ring having a negative area
as calculated by applying the surveyor's formula to the vertices of
the polygon in tile coordinates. In the tile coordinate system (with
the Y axis positive down and X axis positive to the right) this makes
the interior ring's winding order appear counterclockwise.
```