Make lwgeom_subdivide use lwgeom_intersection for clipping
lwgeom_clip_by_rect describes itself as clipping in a "fast but possibly
dirty way." This is not a good fit for lwgeom_subdivide, a primary use of
which is the optimization of point-in-polygon processes. That application
requires that the area covered by the subdivided geometry be the same as
the area covered by the original geometry, which in turn requires that a
robust intersection routine be used.

References #4038
Closes https://github.com/postgis/postgis/pull/281



git-svn-id: http://svn.osgeo.org/postgis/trunk@16676 b70326c6-7e19-0410-871a-916f4a2858ee