Enhance test for topology edge side faces
- Also check non-isolated ones
- Catch universal face containment (only in non-boundary case)

Closes #4830 again

Includes more testcases and updates pre-existing one expectance.

The only missing check now is for dangling edges advertising
the universal face on only ONE side. We cannot easily check
this becase we'd expect such edge to be only contaied by a face
and it will be, even if both sides are in the _same_ face.
This case can only be determined by edge-linking checking
(references #3042)