Add GUS postgis.gdal_vsi_options
For users with out-db rasters in cloud object storage, the
ability to set the config options for the VSI access is needed.
In particular, the authentication keys need to be set for
access to non-public rasters. For example,
```
SET postgis.gdal_vsi_options='AWS_ACCESS_KEY_ID=xxxxxxxxxxxx AWS_SECRET_ACCESS_KEY=yyyyyyyyyyyyyyyyy';
```
This commit also removes the ST_GDALOpenOptions() function,
which was never part of a release, in favour of the GUC approach.