Allow upgrades in presence of views using deprecated functions
Support adding a 'Replaces' comment above functions to improve upgrade

When a function has a line like the following in in the comments
above its signature:

  Replaces ST_Intersection(geometry, geometry) deprecated in 3.1.0

The upgrade script will try to drop the deprecated functions and in case
of failure leave them behind renamed with a _deprecated_by_postgis_<version>
suffix, warns the user about such occurrence and gives her hints on
how to fix.

The postgis_full_version() will report presence of such deprecated
functions as an "incomplete upgrade".

Closes #5033 in 3.1 branch (3.1.5dev)