[winnie] disable path conversion for all scripts
References #5436

At least one winnie job seems to be using build_postgis.sh
instead of regress_postgis.sh, I'm not sure this is intentional, but
this commit should make that one work.

https://winnie.postgis.net/job/PostGIS_2_windows_PGVERSION_winnie/14540/consoleFull

This commit adds a shared file to be sourced by all scripts.
As a bonus, adds `set -e` to the `package_postgis.sh`, making it stop
on first failure.