Fix hard-upgrades failing to skip old st_asgeojson signature
Uses "Replaces" comment instead of _postgis_drop_function_by_identity
function. The latter is only required when the function signature
remains the same but argument names change. These cases are of no
interest for postgis_restore because only the signature (and not the
identity) can be found in PostgreSQL dumps, and so we didn't bother
to implement support for _postgis_drop_function_by_identity in the
script that extracts the skip-list for postgis_restore.

Closes #5623