Apply extension-upgrade-sql-reproducible.patch from trac ticket #5666
This patch is intended to remove a source of non-deterministic build
output, by reading the SOURCE_DATE_EPOCH[1] build-time-indicator
variable and using the resulting value instead of the build-host's
clock as a timestamp in extension upgrade script comments.

The approach here borrows from a similar pattern[2] used in the
'configure.ac' script.

[1] - https://reproducible-builds.org/docs/source-date-epoch/

[2] - https://git.osgeo.org/gitea/postgis/postgis/src/commit/0418cc0ddfda973ba10aa8cc981b07f25a464bdb/configure.ac#L1120

Signed-off-by: jayaddison <jay+osgeo@jp-hosting.net>

Closes #5666 for PostGIS 3.4.3