reference ST_ClipByBox2 in ST_Intersection manual
I have an app using postgis. I was using ST_Intersection to compute the
intersection of geometries with a bounding box and it was super slow.
What I realled needed was ST_ClipByBox2, but I was not aware of that
method. I would have found it easier it had been referenced in the "See
Also" of ST_ClipByBox2